Apply linting fixes from Black and Ruff
- Formatted with Black
- Fixed linting issues with Ruff
- All tests passing (2 files changed)
